bulk Order changes

+++ app/src/main/java/com/delhivery/axle/data/bids/VehicleBidData.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 data class VehicleBidData(
 
     @SerializedName("bid_price") var bidPrice: Double,
-    @SerializedName("truck_capacity")var truckCapacity: Int,
+    @SerializedName("truck_capacity")var truckCapacity: Double,
     @SerializedName("truck_count")var truckCount: Int,
     @SerializedName("vehicle_type")var vehicleType: String,
     @SerializedName("freight_cost")var freightCost: Double =0.0
